New plugin jspsych-cloze
Proposal for a new plugin implementing the Cloze test, well known in the
language teaching literature. Given a specially formatted text, produce
the same text but with select boxes in lieu of particular words.
Participants must select the correct word to complete the text. A score
is calculated at the end. To insert a test word, juste use  the
following syntax: ${}. inside the brackets, separate the options you
wish to present with a "~" (tilde) character. Additionally, precede the
correct choice with an "=" (equal) sign after the tilde.